implement edit functions

+++ include/orderbook.h
@@ -15,18 +15,18 @@ class OrderLevel
 class OrderBook
 {
 private:
-    // bid orders need to match from high -> low
-    std::map<double, std::vector<std::shared_ptr<Order>>> bids;
-    // ask orders match from high -> low
-    std::map<double, std::vector<std::shared_ptr<Order>>> asks;
+    // bid orders need to match prices from high -> low
+    std::map<int, std::vector<std::shared_ptr<Order>>> bids;
+    // ask orders match prices from high -> low
+    std::map<int, std::vector<std::shared_ptr<Order>>> asks;
 
 public:
     bool check_order(OrderSide side, double price);
     bool add_order(std::shared_ptr<Order> order_ptr);
     bool delete_order(std::shared_ptr<Order> order_ptr, int id);
 
-    const std::map<double, std::vector<std::shared_ptr<Order>>> &get_bids();
-    const std::map<double, std::vector<std::shared_ptr<Order>>> &get_asks();
+    const std::map<int, std::vector<std::shared_ptr<Order>>> &get_bids();
+    const std::map<int, std::vector<std::shared_ptr<Order>>> &get_asks();
     void print_orders();
 };
